5.0.0 changes
Reworked the entire project to account for the following:

1. A more modern build system workflow, including a flattened structure
and the use of up-to-date dependencies, such as DartSass instead of
LibSass.
2. The use of SCSS instead of SASS.
3. A more modern SASS approach, such as utilizing @use/@forward instead
of @import, overrideable variables, etc.

These changes will make pulling this project in as a SASS library much
easier, allowing for only pulling in the things you want/need,
customizing every aspect of the output CSS, and more.